requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
requests
